Updated advancemants files for new recipes
I am no happy with how advancements are done at the moment, but
have no knowldge or need to do something about it, so it will stay this
way for now.

Also, there was a gravel.json file I deleted, I have no idea what it was
used for, it seems utterly redundant.